Move useDetectOnlineStatus to join other utils
